refactor: 업로드한 이미지 파일명 파싱 함수 분리
